+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 > Read detailed change logs in https://github.com/sakurawald/fuji/commits/dev/
 
+- [top_chunks] fix: should respect the `top.rows` and `top.columns` options, while display the `Top Chunks` GUI.
+- [top_chunks] fix: should attach the `nearest players` into `chunk scores` while displaying the `Top Chunks` GUI.